fix tests

+++ test/concrete/ob/OrderBook.deposit.t.sol
@@ -19,7 +19,7 @@ contract OrderBookDepositTest is OrderBookExternalMockTest {
     /// Tests that we can deposit some amount and view the new vault balance.
     /// forge-config: default.fuzz.runs = 100
     function testDepositSimple(address depositor, bytes32 vaultId, uint256 amount18) external {
-        amount18 = bound(amount18, 1, type(uint256).max / 10);
+        amount18 = bound(amount18, 1, uint256(int256(type(int224).max)) / 10);
         vm.prank(depositor);
         vm.mockCall(
             address(iToken0),
@@ -84,7 +84,7 @@ contract OrderBookDepositTest is OrderBookExternalMockTest {
     /// Any failure in the deposit should revert the entire transaction.
     /// forge-config: default.fuzz.runs = 100
     function testDepositFail(address depositor, bytes32 vaultId, uint256 amount18) external {
-        amount18 = bound(amount18, 1, type(uint256).max / 10);
+        amount18 = bound(amount18, 1, uint256(int256(type(int224).max)) / 10);
         Float amount = LibDecimalFloat.fromFixedDecimalLosslessPacked(amount18, 18);
 
         // The token contract always reverts when not mocked.
@@ -124,7 +124,7 @@ contract OrderBookDepositTest is OrderBookExternalMockTest {
         vm.assume(actions.length > 0);
         for (uint256 i = 0; i < actions.length; i++) {
             // Deposit amounts must be non-zero.
-            actions[i].amount18 = bound(actions[i].amount18, 1, type(uint256).max / 10);
+            actions[i].amount18 = bound(actions[i].amount18, 1, uint256(int256(type(int224).max)) / 10);
             actions[i].amount = LibDecimalFloat.fromFixedDecimalLosslessPacked(actions[i].amount18, 18);
             // Avoid errors from attempting to etch precompiles.
             vm.assume(uint160(actions[i].token) < 1 || 10 < uint160(actions[i].token));
@@ -179,7 +179,7 @@ contract OrderBookDepositTest is OrderBookExternalMockTest {
     /// Depositing should emit an event with the sender and all deposit details.
     /// forge-config: default.fuzz.runs = 100
     function testDepositEvent(address depositor, bytes32 vaultId, uint256 amount18) external {
-        amount18 = bound(amount18, 1, type(uint256).max / 10);
+        amount18 = bound(amount18, 1, uint256(int256(type(int224).max)) / 10);
         vm.prank(depositor);
         vm.mockCall(
             address(iToken0),
@@ -202,8 +202,8 @@ contract OrderBookDepositTest is OrderBookExternalMockTest {
         bytes32 reVaultId,
         uint256 reAmount18
     ) external {
-        amount18 = bound(amount18, 1, type(uint256).max / 10);
-        reAmount18 = bound(reAmount18, 1, type(uint256).max / 10);
+        amount18 = bound(amount18, 1, uint256(int256(type(int224).max)) / 10);
+        reAmount18 = bound(reAmount18, 1, uint256(int256(type(int224).max)) / 10);
         vm.prank(depositor);
         Reenteroor reenteroor = new Reenteroor();
         Float amount = LibDecimalFloat.fromFixedDecimalLosslessPacked(amount18, 18);
